requirements.txt
pytorch == 1.5.0
torchvision == 0.6.0
numpy == 1.18.1
scikit-learn == 0.22.1
scikit-image == 0.16.2

